+++ middle/middle 11 12 13 14 15.cpp
@@ -1,14 +1,14 @@
 #include "middle.h"
 #include "easy.h"
 
-int itc_second_simple_max_num(long long num) {
+long long  itc_second_simple_max_num(long long num) {
 	num = itc_abs(num);
 	
-	int max = -1, max1 = -1, num1 = num, i = 0, i2 = 0, idx = 0;
+	long long  max = -1, max1 = -1, num1 = num, i = 0, i2 = 0, idx = 0;
 
 	while (num > 0) {
 		i++;
-		int d = num % 10;
+		long long  d = num % 10;
 		if (d > max) {
 			max = d;
 			idx = i;
@@ -17,7 +17,7 @@ int itc_second_simple_max_num(long long num) {
 	}
 	while (num1 > 0) {
 		i2++;
-		int d = num1 % 10;
+		long long  d = num1 % 10;
 		if (i2 != idx && max1 < d) max1 = d;
 		num1 /= 10;
 	}
@@ -29,7 +29,7 @@ long long itc_bin_num(long long number) {
 	
 	long long a = 0, fd = -1;
 	while (number > 0) {
-		int t = number % 2;
+		long long  t = number % 2;
 		if (fd == -1) fd = t;
 		a = a * 10 + t;
 		number /= 2;
@@ -47,7 +47,7 @@ long long itc_oct_num(long long number) {
 	number = itc_abs(number);
 	long long a = 0, fd = -1;
 	while (number > 0) {
-		int t = number % 8;
+		long long  t = number % 8;
 		if (fd == -1) fd = t;
 		a = a * 10 + t;
 		number /= 8;
@@ -61,21 +61,21 @@ long long itc_oct_num(long long number) {
 	if (fd == 0) m *= 10;
 	return m;
 }
-int itc_rev_bin_num(long long number) {
+long long  itc_rev_bin_num(long long number) {
 	number = itc_abs(number);
 	long long a = 0, len = itc_len_num(number);
 	for (long long i = 0; number > 0; i++) {
-		int t = number % 10;
+		long long  t = number % 10;
 		a += t * itc_pow(2, i);
 		number /= 10;
 	}
 	return a;
 }
-int itc_rev_oct_num(long long number) {
+long long  itc_rev_oct_num(long long number) {
 	number = itc_abs(number);
 	long long a = 0, len = itc_len_num(number);
 	for (long long i = 0; number > 0; i++) {
-		int t = number % 10;
+		long long  t = number % 10;
 		a += t * itc_pow(8, i);
 		number /= 10;
 	}
